refactor(frontend): extract chat composer
