+++ packages/workshop-frontend/src/features/chat/attachmentFormatting.ts
@@ -1,0 +1,6 @@
+export function formatAttachmentSize(size: number | undefined): string | null {
+  if (size === undefined) return null;
+  if (size < 1024) return `${size} B`;
+  if (size < 1024 * 1024) return `${(size / 1024).toFixed(1)} KB`;
+  return `${(size / (1024 * 1024)).toFixed(1)} MB`;
+}
